Add some example copies of the session model for each app
These are the result of `req.sessionModel.toJSON()` at the end of an application, with things which are known to be redundant removed.

In the case of S5, this is the amalgamation of a couple of journeys, since some fields only exist if some questions are answered in a specific way.